+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/trigger/GroovyTriggerMaker.java
@@ -9,8 +9,6 @@ import com.dwarfeng.fdr.stack.exception.TriggerMakeException;
 import com.dwarfeng.fdr.stack.handler.Trigger;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import groovy.lang.GroovyClassLoader;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.ApplicationContext;
@@ -79,8 +77,6 @@ public class GroovyTriggerMaker implements TriggerMaker {
     @Component
     @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
     public static class GroovyTrigger implements Trigger {
-
-        private static final Logger LOGGER = LoggerFactory.getLogger(GroovyTrigger.class);
 
         private LongIdKey pointKey;
         private LongIdKey triggerInfoKey;